update types, refactor
