+++ src/helper/mutations/helper.ts
@@ -22,7 +22,7 @@ import { RefObject } from 'react'
 import { csrReload } from '../csr-reload'
 import { showToastNotice } from '../show-toast-notice'
 import { triggerSentry } from '../trigger-sentry'
-import { SetEntityInputTypes } from './use-set-entity-mutation'
+import { SetEntityInputTypes } from './use-set-entity-mutation/types'
 import { endpoint } from '@/api/endpoint'
 import { AuthenticationPayload } from '@/auth/auth-provider'
